% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{alignAS}
\alias{alignAS}
\title{Align a highlighted region's AS clauses within a SQL statement.}
\usage{
alignAS()
}
\value{
Aligns the AS clauses in a SQL statement within a highlighted region.
}
\description{
Align a highlighted region's AS clauses within a SQL statement.
}
