% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{alignColon}
\alias{alignColon}
\title{Align a highlighted region's : separators.}
\usage{
alignColon()
}
\value{
Aligns the : delimiter found in YAML within a highlighted region.
}
\description{
Align a highlighted region's : separators.
}
