% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{alignCustom}
\alias{alignCustom}
\title{Align a given regular expression.}
\usage{
alignCustom()
}
\value{
Aligns regular expression within a highlighted region.
}
\description{
Align a given regular expression.
}
