% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{alignRegex}
\alias{alignRegex}
\title{Align a given regular expression.}
\usage{
alignRegex(x, ...)
}
\arguments{
\item{x}{Custom input; a regular expression.}

\item{...}{Additional arguments passed along to \code{regexec()}}
}
\value{
Aligns regular expression within a highlighted region.
}
\description{
Align a given regular expression.
}
