% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{alignTilde}
\alias{alignTilde}
\title{Aligns the tilder '~' within a highlighted area.}
\usage{
alignTilde()
}
\value{
Aligns the tilde (\code{~}) within a highlighted region.
}
\description{
Aligns the tilder '~' within a highlighted area.
}
